Ansible and shell scripts to automate team change
buildAllTeams.sh unpacks, builds, and repacks each team's archive file in /home/swarmie/physicalTeams2017

populalateArchives.yml is an ansible play for copying team archives to all swarmies, and copying and running the buildAllTeams.sh script.  This file is intended to be run after pulling and archiving all teams' code on the local machine.

loadTeam.yml is a ansible play to be run during the Swarmathon competition.  This play prompts for the team's code as input, then it prepares the team's code to be run on each swarmie.  This is to be done prior to running the launch script for each robot at the beginning of each round.